--- api/pkg/identity/policy.csv
@@ -1,53 +1,0 @@
-p,role_admin,/apisix/admin/ssl,POST
-p,role_admin,/apisix/admin/services,POST
-p,role_admin,/apisix/admin/stream_routes,POST
-p,role_admin,/apisix/admin/system_config,POST
-p,role_admin,/apisix/admin/upstreams,POST
-p,role_admin,/apisix/admin/plugin_configs,POST
-p,role_admin,/apisix/admin/proto,POST
-p,role_admin,/apisix/admin/routes,POST
-p,role_admin,/apisix/admin/import/routes,POST
-p,role_admin,/apisix/admin/ssl,PUT
-p,role_admin,/apisix/admin/ssl/:id,PUT
-p,role_admin,/apisix/admin/services,PUT
-p,role_admin,/apisix/admin/services/:id,PUT
-p,role_admin,/apisix/admin/stream_routes,PUT
-p,role_admin,/apisix/admin/stream_routes/:id,PUT
-p,role_admin,/apisix/admin/system_config,PUT
-p,role_admin,/apisix/admin/plugin_configs,PUT
-p,role_admin,/apisix/admin/plugin_configs/:id,PUT
-p,role_admin,/apisix/admin/proto,PUT
-p,role_admin,/apisix/admin/proto/:id,PUT
-p,role_admin,/apisix/admin/routes,PUT
-p,role_admin,/apisix/admin/routes/:id,PUT
-p,role_admin,/apisix/admin/consumers,PUT
-p,role_admin,/apisix/admin/consumers/:username,PUT
-p,role_admin,/apisix/admin/upstreams,PUT
-p,role_admin,/apisix/admin/upstreams/:id,PUT
-p,role_admin,/apisix/admin/global_rules,PUT
-p,role_admin,/apisix/admin/global_rules/:id,PUT
-p,role_admin,/apisix/admin/ssl/:ids,DELETE
-p,role_admin,/apisix/admin/services/:ids,DELETE
-p,role_admin,/apisix/admin/stream_routes/:ids,DELETE
-p,role_admin,/apisix/admin/system_config/:config_name,DELETE
-p,role_admin,/apisix/admin/plugin_configs/:ids,DELETE
-p,role_admin,/apisix/admin/proto/:ids,DELETE
-p,role_admin,/apisix/admin/routes/:ids,DELETE
-p,role_admin,/apisix/admin/consumers/:usernames,DELETE
-p,role_admin,/apisix/admin/upstreams/:ids,DELETE
-p,role_admin,/apisix/admin/global_rules/:id,DELETE
-p,role_admin,/apisix/admin/ssl/:id,PATCH
-p,role_admin,/apisix/admin/ssl/:id/*path,PATCH
-p,role_admin,/apisix/admin/services/:id,PATCH
-p,role_admin,/apisix/admin/services/:id/*path,PATCH
-p,role_admin,/apisix/admin/plugin_configs/:id,PATCH
-p,role_admin,/apisix/admin/plugin_configs/:id/*path,PATCH
-p,role_admin,/apisix/admin/proto/:id,PATCH
-p,role_admin,/apisix/admin/proto/:id/*path,PATCH
-p,role_admin,/apisix/admin/routes/:id,PATCH
-p,role_admin,/apisix/admin/routes/:id/*path,PATCH
-p,role_admin,/apisix/admin/upstreams/:id,PATCH
-p,role_admin,/apisix/admin/upstreams/:id/*path,PATCH
-p,role_admin,/apisix/admin/global_rules/:id,PATCH
-p,role_admin,/apisix/admin/global_rules/:id/*path,PATCH
-g,user_admin,role_admin
